Change type? int

+++ src/SalesManagerRuner.java
@@ -2,8 +2,8 @@ import netology.SalesManager;
 
 public class SalesManagerRuner {
     public static void main(String[] args) {
-        long result;
-        long[] sales = {1, 3, 5, 7};
+        int result;
+        int[] sales = {1, 3, 5, 7};
         //int result;
         //int[] sales = {1, 3, 5, 7};
         SalesManager SM = new SalesManager(sales);
